december01: Move to return long
Doesn't matter for now but eventually that'll bite me :)

+++ src/main/java/com/gordonreid/adventofcode2023/december01/Part1.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 final class Part1 {
 
-    static int run(List<String> input) {
+    static long run(List<String> input) {
         return input.stream()
                 .map(Common::stripNonDigits)
                 .map(Common::getFirstAndLastDigitAsString)
